renamed multiimagingextractor --> FrameConcatenatedImagingExtractor
